파일 저장 기능

+++ EngineBase/EngineSerializer.cpp
@@ -19,7 +19,7 @@ void UEngineSerializer::operator>> (ISerializObject& _Data)
 	_Data.DeSerialize(*this);
 }
 
-void UEngineSerializer::Write(void* _Data, size_t _Size)
+void UEngineSerializer::Write(void* _Data, int _Size)
 {
 	if (Data.size() <= WriteOffset + _Size)
 		Data.resize(Data.size() * 2 + _Size);
@@ -28,7 +28,7 @@ void UEngineSerializer::Write(void* _Data, size_t _Size)
 	WriteOffset += _Size;
 }
 
-void UEngineSerializer::Read(void* _Data, size_t _Size)
+void UEngineSerializer::Read(void* _Data, int _Size)
 {
 	memcpy_s(_Data, _Size, &Data[ReadOffset], _Size);
 	ReadOffset += _Size;
